🔧 Fix: modify install script for broader compatibility (#59)
* 🔧 Fix: modify install script for broader compatibility

I updated the bash install script so that it would work in my Fedora 41 environment. It still likely needs testing in other environments, but the key differences will be outlined in the PR related to this commit. The key differences include permissions handling, tmp folder checking, and more robust error handling.

* 🔧 Fix: fixed error handling on /tmp/ for install.sh

This commit adds additional error handling, ensuring that the script exits gracefully even when /tmp/ dir clearance is not allowed due to permissions errors. This is particularly relevant  in distros using SELinux which prevents users from deleting directories they don't own, even with sudo. Now, clean up is handled or user is handed an error message and the script continues to run as intended.