:wrench: fix: Revamped `help <cmd>` and `--help` flag  (#69)
* ✨ Feat: Modify help command to display version

Version was previously omitted from the list of available commands. Having it explicitly listed is an ease-of-use thing for developers.

* 🚚 [chore] Rename commanderr.go to commander.go

This commit fixes a typo where commander had two trailing r's instead of a single one. This is a purely cosmetic fix and shouldn't meaningfully affect the functionality of the manuscript-cli.

* 🎨 [chore] group initialized cmds into logical groups

Reorganizes the body of the init() function in commander.go to put a logical order where commands go so that it is more easily followed. Also adds comments to better describe what's happening.

* 📝 [docs] Add Help Context to all Cmds

This commit adds Long text and Example text to each command which more adequately describes their functions and capabilities.

* ♻️ [refactor] Reorganize help system using cobra templates

This change is designed to make the help system more maintainable and consistent. I replaced custom help function with native cobra templates. I created a separate `help_template.go` to keep separation of concerns. And I removed the fragile manual command listing and formatting.

* 🎨 [chore] Beautify help output

This commit makes help output more concise and more beautiful - adding emojis and simplifying  some of the text in help so as to be more accurate.

* 💄 [fix] Disable alphabetical command sorting

This commit ensures that commands are displayed by the help message in the logical order they were added as opposed to in alphabetical order sorted by Cobra.