:sparkles: feat: enhance cli version cmd (#76)
* ✨ Feat: introduce new version --verbose flag

This commit introduces a `--verbose`/`-v` flag for the version command and the top-level logic for the version command to handle verbosity requests.

* ✨ Feat: add version verbose logic

This commit adds `version_manuscript.go` as a new file to the repo. Within, new structs are defined for handling version information for manuscript-core, docker, docker desktop,as well as the build time environment of rust, and go. Additionally, versioning information is  pretty-printed to the console & emojified!

* 🔖 [chore] bump version number to 1.1.0

The CLI was still displaying the old version number (1.0.4) ; this commit bumps the version var in the commander.go package so that the version command will display the proper version.

* 💡 [chore] update build information about ldflags

There was an error previously. Wording  did not contain the correct path for the `go build`instructions in the source code comments of `version_manuscript.go`