:hammer: feat: add Makefile for Unified Build System (#83)
* 🔨 Feat: add Makefile for CLI & GUI

This commit adds a makefile which can be used to install either the CLI, the GUI, or both. It also allows for a single interface to interact with both rust and go commands respectively.

* 🔨 Fix: remove build dependency for make install

Removed where install automatically built both binaries. Separate install and build steps. Running build commands with sudo causes problems because sudo has a different environment

* 🔨 Feat: add sophisticated error handling in Makefile

This adds significant error handling in each step of the Makefile. Install verifies that binaries are in the right location. Build verifies that binaries are created.  Permissions are checked to prevent permissions errorsand most failures are handled gracefully. This can also help with debugging any issues during builds by giving this detailed user feedback