:bug: fix: eliminate zombie jobs (#96)
* 🎨 [chore] create chainEndpoint constant

This commit moves the hard-coded network chains endpoint into a `chainEndpoint` constant for better visibility and configurability.

* ✨ Feat: add manuscript state machine

Added a Manuscript State Detector which can be instanciated in the ListJobs() function for sake of identifying manuscript state with accuracy and finality.  The state machine executes a series of elaborate looks at container status and flink logs to  update to one  of five states: RUNNING, INITIALIZING, FAILED, STOPPED, and UNKNOWN (rare). This is the grounds for more advanced manuscript detection from the CLI.

* ♻️ [chore] refactor listjobs to accomodate state machine

This commit adds new logic to jobs_manuscipt that allows for the detection of manuscripts by using `manuscript_state.go`

* 💡 [chore] add comments to state machine

Recognizing that there is a lot of complicated logic in the state detector, I added comments explaining its general operation throughout.

* 🐛 Fix: enhance manuscript state detection

Within the DetectState method, manuscripts previously defaulted to initializing state, despite surviving all previous checks. Now, if a manuscript survives all checks it will display as `RUNNING` in the jobs list.

* 🦺 [chore] add graphQL check to state detector

This adds as an additional state check - the checkGraphQLEndpoint function which hits the local endpoint to determine healthiness and ensure it is live before returning RUNNING. If it is determined not to be live, but survive all other checks, the manuscript will return INITIALIZING status.

* 📝 [chore] add additional help documentation for cli `list` cmd

This commit adds additional status indicators to the long help text of the list command.

* ✨ [feat] add alias `ls` to cli `list` command

This commit adds an alias, `ls` to the list command.

* 🦺 [chore] add guard rails for cli init

This commit adds additional guard rails to fix broken functionality of `init` command that was affecting the jobs list. Previously, a person could init and partially overwrite an existing manuscript. Now, extensive checks are in place to prevent that from happening: manuscript is looked for in the directory of the manuscript path, the manuscript is already running, the manuscript is in the config file.

* ✨ [feat] enhance `manuscript-cli list` cmd

This commit adds new functionality to manuscript-cli list. It now accepts a directory as an argument. If no directory is specified it looks in the default directory for users. It also shows STOPPED jobs by looking at the manuscript yamls in the subdirectories of the specified directory and showing ALL manuscripts. This makes it more informative and usable.