✨ feat: add non-interactive manuscript `init` CLI command (#110)
* ✨ Feat: add non-interactive CLI `init` logic

This commit adds the basic functionality to generate a manuscript non-interactively to `init_manuscript.go.` It is a preliminary action and violates some DRY
principles for ease of use and simple implementation. In future versions, it will be wise to
refactor both `init` paths to use common helper funcs.

* ✨ Feat: add `init` non-interactive and flags to CLI

This commit adds flags and pre-run validation to the non-interactive version of the `init` command. It greatly expands the functionality of the command to be able to take arguments directly.

* 🐛 Fix: manuscripts create in proper directory

This commit fixes the `init` non-interactive to ensure manuscripts are created in the proper place.